refactor(step-bar): コンポーネントの呼び出し方を統一
